requirements.txt
torch==1.12.1+cu113 --extra-index-url https://download.pytorch.org/whl/cu113
torchvision==0.13.1+cu113 --extra-index-url https://download.pytorch.org/whl/cu113
torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu113
farm-haystack @ git+https://github.com/deepset-ai/haystack.git@54bf7ad3439a9c481cb8d32ca720612a1341deb8


